+++ frontend/src/App.tsx
@@ -2,7 +2,7 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import NavBar from "./components/NavBar.tsx";
 import {Route, Routes} from "react-router-dom";
 import ProductList from "./pages/products/ProductList.tsx";
-import MarketList from "./pages/products/MarketList.tsx";
+import MarketList from "./pages/market/MarketList.tsx";
 
 export default function App() {
